+++ backends/database/geohub-database.sql
@@ -373,3 +373,98 @@ CREATE TABLE IF NOT EXISTS geohub.stac_collection_product
 
 COMMENT ON TABLE geohub.stac_collection_product
     IS 'This is the table to manage linking a product to stac collection.';
+
+--------------------------
+-- Storymap tables
+--------------------------
+
+CREATE TABLE IF NOT EXISTS geohub.storymap
+(
+    id uuid NOT NULL,
+    title character varying NOT NULL,
+    logo bytea,
+    subtitle character varying,
+    byline character varying,
+    footer character varying,
+    template_id character varying NOT NULL,
+    style_id integer,
+    base_style_id character varying,
+    createdat timestamp with time zone NOT NULL,
+    created_user character varying NOT NULL,
+    updatedat timestamp with time zone,
+    updated_user character varying,
+    CONSTRAINT storymap_pkey PRIMARY KEY (id)
+);
+
+CREATE TABLE IF NOT EXISTS geohub.storymap_chapter
+(
+    id uuid NOT NULL,
+    title character varying NOT NULL,
+    description character varying NOT NULL,
+    image bytea,
+    image_alignment character varying,
+    alignment character varying NOT NULL,
+    map_interactive boolean NOT NULL DEFAULT false,
+    map_navigation_position character varying NOT NULL,
+    map_animation character varying NOT NULL,
+    rotate_animation boolean NOT NULL DEFAULT false,
+    spinglobe boolean NOT NULL DEFAULT false,
+    hidden boolean NOT NULL DEFAULT false,
+    center geometry(Point,4326) NOT NULL,
+    zoom double precision NOT NULL,
+    bearing double precision NOT NULL DEFAULT 0,
+    pitch double precision NOT NULL DEFAULT 0,
+    style_id integer,
+    base_style_id character varying,
+    on_chapter_enter jsonb,
+    on_chapter_exit jsonb,
+    createdat timestamp with time zone NOT NULL,
+    created_user character varying NOT NULL,
+    updatedat timestamp with time zone,
+    updated_user character varying,
+    CONSTRAINT storymap_chapter_pkey PRIMARY KEY (id)
+);
+
+CREATE TABLE IF NOT EXISTS geohub.storymap_chapters
+(
+    storymap_id uuid NOT NULL,
+    chapter_id uuid NOT NULL,
+    sequence integer NOT NULL,
+    CONSTRAINT storymap_chapters_pkey PRIMARY KEY (storymap_id, chapter_id),
+    CONSTRAINT fk_storymap_to_storymap_chapters FOREIGN KEY (storymap_id)
+        REFERENCES geohub.storymap (id) MATCH SIMPLE
+        ON UPDATE NO ACTION
+        ON DELETE CASCADE,
+    CONSTRAINT fk_storymap_to_storymap_chapter FOREIGN KEY (chapter_id)
+        REFERENCES geohub.storymap_chapter (id) MATCH SIMPLE
+        ON UPDATE NO ACTION
+        ON DELETE CASCADE
+);
+
+CREATE TABLE IF NOT EXISTS geohub.storymap_favourite
+(
+    storymap_id uuid NOT NULL,
+    user_email character varying(100) NOT NULL,
+    savedat timestamp with time zone NOT NULL,
+    CONSTRAINT storymap_favourite_pkey PRIMARY KEY (storymap_id, user_email),
+    CONSTRAINT "FK_storymap_TO_storymap_favourite" FOREIGN KEY (storymap_id)
+        REFERENCES geohub.storymap (id) MATCH SIMPLE
+        ON UPDATE NO ACTION
+        ON DELETE CASCADE
+        NOT VALID
+);
+
+CREATE TABLE IF NOT EXISTS geohub.storymap_permission
+(
+    storymap_id uuid NOT NULL,
+    user_email character varying(100) NOT NULL,
+    permission smallint NOT NULL DEFAULT 1,
+    createdat timestamp with time zone NOT NULL DEFAULT now(),
+    updatedat timestamp with time zone,
+    CONSTRAINT storymap_permission_pkey PRIMARY KEY (storymap_id, user_email),
+    CONSTRAINT "FK_storymap_TO_storymap_permission" FOREIGN KEY (storymap_id)
+        REFERENCES geohub.storymap (id) MATCH SIMPLE
+        ON UPDATE NO ACTION
+        ON DELETE CASCADE
+        NOT VALID
+);