+++ sites/geohub/src/components/pages/data/datasets/DefaultStyleEditor.svelte
@@ -177,14 +177,16 @@
 		$map.resize();
 	};
 
-	$: layerType, handleLayerSelected();
-
 	const handleLayerSelected = async () => {
 		if (!$map) return;
 		try {
 			isLoading = true;
 
 			const style = $map.getStyle();
+
+			if (!is_raster) {
+				sourceId = feature.properties.id;
+			}
 
 			if (sourceId && $map.getSource(sourceId)) {
 				const layers = style.layers.filter(
@@ -458,7 +460,11 @@
 								</div>
 							{/if}
 							<div class="mt-2">
-								<LayerTypeSwitch bind:layer={selectedVectorLayer} bind:layerType />
+								<LayerTypeSwitch
+									bind:layer={selectedVectorLayer}
+									bind:layerType
+									on:change={handleLayerSelected}
+								/>
 							</div>
 						</div>
 					{/if}