+++ sites/geohub/src/lib/server/helpers/createDatasetLinks.ts
@@ -53,6 +53,11 @@ export const createDatasetLinks = async (
 			type: 'image/webp',
 			href: `${origin}/api/datasets/${feature.properties.id}/preview/auto/{width}x{height}.webp`
 		});
+		feature.properties.links.push({
+			rel: 'stylejson',
+			type: 'application/json',
+			href: `${origin}/api/datasets/${feature.properties.id}/preview/style.json`
+		});
 	} else if (type?.value === 'stac') {
 		const stacType = tags?.find((tag) => tag.key === 'stacType')?.value;
 		const product = tags?.find((t) => t.key === 'product')?.value;
@@ -85,6 +90,11 @@ export const createDatasetLinks = async (
 					)}&expression=${encodeURIComponent(
 						expression
 					)}&asset_as_band=true&scale=1&bidx=1&resampling=nearest&return_mask=true`
+				});
+				feature.properties.links.push({
+					rel: 'stac',
+					type: 'application/json',
+					href: `${titilerUrl}`
 				});
 				feature.properties.links.push({
 					rel: 'tilejson',
@@ -193,6 +203,11 @@ export const createDatasetLinks = async (
 			rel: 'preview',
 			type: 'image/webp',
 			href: `${origin}/api/datasets/${feature.properties.id}/preview/auto/{width}x{height}.webp`
+		});
+		feature.properties.links.push({
+			rel: 'stylejson',
+			type: 'application/json',
+			href: `${origin}/api/datasets/${feature.properties.id}/preview/style.json`
 		});
 
 		if (is_raster) {