+++ sites/geohub/src/lib/server/helpers/index.ts
@@ -29,3 +29,4 @@ export * from './getDatasetStats';
 export * from './stac';
 export * from './searchUsersByEmail';
 export * from './resolveSpriteUrl';
+export * from './isGeoHubBlobStorage';